==> notebooks/1_data2csv.ipynb <==
# Visualize data
This notebooks transforms data pickle files to csv for better visualization of the data structure.

import sys
import os
import re
import pickle as pkl
import pandas as pd

DATA_DIR='../data/bias_data'
%cd {DATA_DIR}

def get_file():
    for dir_ in os.listdir('.'):
        if not os.path.isdir(dir_) or dir_ == 'model_vocab':
            continue
        for file in os.listdir(dir_):
            if re.match(r'^.*\.pkl$', file):
                yield os.path.join(dir_, file)

def load_df(path):
    with open(path, "rb") as f:
        obj = pkl.load(f)

    df = pd.DataFrame(obj)
    #df.to_csv(r'file.csv')
    return df

## Show example dataset
Here we show an example of the model captions.

Each row of the data corresponds to an image in the COCO dataset. It contains the id of the image (`img_id`), the caption produced by the model (`pred`), and the ground truth annotations of the protected attributes gender and race (`bb_gender` and `bb_skin`, respectively).

df = load_df('Show-Tell/gender_val_st10_cap_mw_entries.pkl')
df

df.columns

## Make a copy of the data in csv format

for src in get_file():
    print(src)
    df = load_df(src)
    dst = src.replace('.pkl', '.csv')
    df.to_csv(dst)

## Show example vocabulary
Here we show an example of the vocabulary used for a model.

path = 'model_vocab/fc_vocab.pkl'
with open(path, "rb") as f:
    obj = pkl.load(f)
print(obj)
