==> notebooks/2_age_models_intersection.ipynb <==
# Select dataset for age
In this notebook we select the images that will be used for the age bias analysis. Our experiments use the models NIC, SAT, NIC_PLUS and NIC_EQUALIZER. The datasets for each model are separated into different files. We decided to use the intersection between the images selected for each of the models, which still keeps enough images for our experiments.

import os
import re
import pickle
import pandas as pd

DATA_DIR = '../data/bias_data'

human_file = 'Human_Ann/gender_obj_cap_mw_entries.pkl'
model_files = {
    'oscar': 'Oscar/gender_val_oscar_cap_mw_entries.pkl',
    'sat': 'Show-Attend-Tell/gender_val_sat_cap_mw_entries.pkl',
    'nicplus': 'Woman-Snowboard/gender_val_baselineft_cap_mw_entries.pkl',
    'niceq': 'Woman-Snowboard/gender_val_snowboard_cap_mw_entries.pkl',
}

## Human captions dataset

human_path = os.path.join(DATA_DIR, human_file)
main_df = pd.DataFrame(pd.read_pickle(human_path))[['img_id', 'caption_list']]
main_df

## Model captions datasets
All the datasets happened to have captions for the same images, so the intersection remains as the whole set.

for model_name, model_file in model_files.items():
    model_path = os.path.join(DATA_DIR, model_file)
    df = pd.DataFrame(pd.read_pickle(model_path))[['img_id','pred']]
    df = df.rename(columns={'pred': f'pred_{model_name}'})
    print(model_name, len(df))
    main_df = pd.merge(main_df, df, how='inner', on=['img_id'])
main_df = main_df.dropna()
main_df

## Save captions
The file contains all captions from all models and human.

main_df.set_index('img_id')

OUT_FILE = '../res/IntersectionSAT-OSCAR-NICPL-NICEQ.csv'
main_df.set_index('img_id').to_csv(OUT_FILE, index_label='img_id')
