==> notebooks/3_prepare_coco.ipynb <==
# Prepare COCO
The COCO dataset contains separate train, test and validation sets. The images have unique IDs.

This notebook combines and renames the images in the different datasets to the format `<img_id>.jpg`.

Then, a reduced version of COCO is created containing only the images that are relevant for the task.

import os
import shutil
from tqdm import tqdm
import pandas as pd

The dataset folders `train2014`, `test2014` and `valid2014` must be [downloaded](https://cocodataset.org/#download), extracted from zipped file and placed in `res/mscoco/`.

RES_DIR = '../res'

%cd {RES_DIR}
os.listdir()

datasets = ['train2014', 'val2014', 'test2014']

idss = []
for ds in datasets:
    ids = []
    ds_path = os.path.join('mscoco', ds)
    if not os.path.exists(ds_path):
        continue
    for filename in os.listdir(ds_path):
        id_ = int(filename.rsplit('_', maxsplit=1)[1].split('.')[0])
        ids.append(id_)
    print(ds, len(os.listdir(ds_path)), len(set(ids)))
    idss.extend(ids)
print('total', len(idss), len(set(idss)))

## Merge datasets in a single folder
Files can be moved or copied.

dst_dir = 'merged_coco'
if not os.path.exists(dst_dir):
    os.makedirs(dst_dir)

for ds in datasets:
    ds_path = os.path.join('mscoco', ds)
    if not os.path.exists(ds_path):
        continue
    for src_file in tqdm(os.listdir(ds_path), desc=ds):
        dst_file = src_file.rsplit('_', maxsplit=1)[1]
        src = os.path.join(ds_path, src_file)
        dst = os.path.join(dst_dir, dst_file)
        
        shutil.copyfile(src, dst)  # copy if you want to keep the previous dataset
        # shutil.move(src, dst)  # move to save space

print(dst_dir, len(os.listdir(dst_dir)))

### Remove previous dataset folders

# for ds in datasets:
#     shutil.rmtree(ds)

## Reduce the COCO dataset for our task
We will just need the images for which we have annotations. We can create a reduced COCO dataset only keeping those images.

df = pd.read_csv('IntersectionSAT-OSCAR-NICPL-NICEQ.csv', index_col='img_id')
df

src_dir = 'merged_coco'
dst_dir = f'reduced_coco_{len(df)}'

if not os.path.exists(dst_dir):
    os.makedirs(dst_dir)

df.index.values

for filename in tqdm(os.listdir(src_dir)):
    id_ = int(filename.split('.')[0])

    if id_ in df.index.values:
        src = os.path.join(src_dir, filename)
        dst = os.path.join(dst_dir, filename)
        
        shutil.copyfile(src, dst)  # copy if you want to keep the previous dataset
        # shutil.move(src, dst)  # move to save space

print(dst_dir, len(os.listdir(dst_dir)))

### Remove temporal `merged_coco` folder

# shutil.rmtree(src_dir)
