==> notebooks/5_prepare_hand_annot.ipynb <==
# Prepare hand annotations dataset
The hand annotations data is reformatted to adapt it to the requirements of the code.

- Rename columns
- Remove exceeding annotations to create balance dataset

NOTE: Creating a balanced dataset is not necessary since the code will balance the data if it is specified.

import os
import random
import pandas as pd

RES_DIR = '../res'

%cd {RES_DIR}

REDUCE_BALANCED = False

SRC_CSV = 'annotations/annotations_merged.csv'

if REDUCE_BALANCED:
    DST_CSV = 'annotations/annotations_merged_renamed_balanced.csv'
else:
    DST_CSV = 'annotations/annotations_merged_renamed.csv'

## Read csv

df = pd.read_csv(SRC_CSV, index_col='id')
df

## Reduce (optional) -> balanced ds

df.label.value_counts()

old_idx = df[df.label == 'old'].index
young_idx = df[df.label == 'young'].index
print(old_idx)
print(young_idx)

limiting_idx = young_idx if len(young_idx) < len(old_idx) else old_idx
excess_idx = old_idx if len(young_idx) < len(old_idx) else young_idx
excess_idx

print('limiting set size:', len(limiting_idx))
print('target balanced ds size:', len(limiting_idx)*2)

reduced_idx = excess_idx[0:len(limiting_idx)]
reduced_idx

df_balanced = df.loc[limiting_idx.union(reduced_idx)]
df_balanced

df2 = df_balanced if REDUCE_BALANCED else df
len(df2)

# Rename columns

df_rn = df2.rename(columns={'label': 'bb_age'})
df_rn.index.name = 'img_id'
df_rn

## Save csv

overwrite = True

if os.path.exists(DST_CSV):
    while True:
        ans = str(input('A csv file already exists. Overwrite? [y]/n: ')).lower()
        if ans == 'y' or ans == 'yes':
            overwrite = True
            break
        elif ans == 'n' or ans == 'no':
            overwrite = False
            break

if overwrite:
    if not os.path.exists(os.path.dirname(DST_CSV)):
        os.makedirs(os.path.dirname(DST_CSV))
    df_rn.to_csv(DST_CSV, index_label='img_id')
    print(f'Saved at {DST_CSV}')
