==> notebooks/6_create_age_dataset.ipynb <==
# Create age dataset

We want to combine the combined lables for age for the images and their captions provided by human annotators, SAT and Oscar

import os
import pandas as pd
import numpy as np
import pickle

Start by loading both sets of data

file_age = '../res/annotations/annotations_merged_renamed.csv'
file_captions = '../res/IntersectionSAT-OSCAR-NICPL-NICEQ.csv'

df_age = pd.read_csv(file_age, index_col='img_id')
df_age

df_captions = pd.read_csv(file_captions, index_col='img_id')
df_captions

df_combined = pd.concat([df_captions, df_age], axis = 1).dropna()
df_combined

# reset_index to match the format of the input data

df_combined.reset_index().to_csv('../data/age_data/age_dataset.csv')

# Redefine vocabulary \[Optional\]
In this section we provide the code to redefine the model vocabulary.

import ast
from collections import Counter

tokens = []

for row in df_combined['caption_list']:
    for cap in ast.literal_eval(row):
        tokens.extend(cap.lower().split())
        
human_most_common = Counter(tokens).most_common()

vocab_human = dict(human_most_common)
print('human', len(vocab_human))
vocab_human

def get_model_vocab(col):
    tokens = []

    for cap in df_combined[col]:
        tokens.extend(cap.lower().split())

    vocab = dict(Counter(tokens).most_common())
    print(col, len(vocab))
    return vocab

vocab_sat = get_model_vocab('pred_sat')
human_vocab_sat = dict(human_most_common[:len(vocab_sat)])

vocab_oscar = get_model_vocab('pred_oscar')
human_vocab_oscar = dict(human_most_common[:len(vocab_oscar)])

vocab_nicplus = get_model_vocab('pred_nicplus')
human_vocab_nicplus = dict(human_most_common[:len(vocab_nicplus)])

vocab_niceq = get_model_vocab('pred_niceq')
human_vocab_niceq = dict(human_most_common[:len(vocab_niceq)])

Save the vocabularies overwriting existing files:

def save_vocabs():

    with open('../data/age_data/model_vocab/sat_vocab.pkl', 'wb') as f:
        pickle.dump(human_vocab_sat, f)
    with open('../data/age_data/model_vocab/oscar_vocab.pkl', 'wb') as f:
        pickle.dump(human_vocab_oscar, f)
    with open('../data/age_data/model_vocab/nic_plus_vocab.pkl', 'wb') as f:
        pickle.dump(human_vocab_nicplus, f)
    with open('../data/age_data/model_vocab/nic_equalizer_vocab.pkl', 'wb') as f:
        pickle.dump(human_vocab_niceq, f)
        
#save_vocabs()  # uncomment this line to overwrite the vocab files that may exists
