==> notebooks/7_csv_pkl_conversion.ipynb <==
# csv-pkl conversion
This notebooks provides util functions to transform `csv` files to `pkl` files and vice versa.

import os
import io
import ast
import re
import pandas as pd
import pickle

def create_dirs(path):
    dir_ = os.path.dirname(path)
    if not os.path.exists(dir_) or not os.path.isdir(dir_):
        os.makedirs(dir_)

## csv2pkl

def csv2pkl(src, dst=None, parse=True):
    if dst is None:
        dst = re.sub(r'\.csv$', '.pkl', src)
    
    # Loading
    df = pd.read_csv(src)
    
    # Parse
    if parse:
        for col in df.columns:
            try:
                df[col] = df[col].apply(lambda x: ast.literal_eval(x))
            except:
                pass
    
    # Overwrite?
    overwrite = True

    if os.path.exists(dst):
        while True:
            ans = str(input('A pkl file already exists. Overwrite? [y]/n: ')).lower()
            if ans == 'y' or ans == 'yes':
                overwrite = True
                break
            elif ans == 'n' or ans == 'no':
                overwrite = False
                break

    # Saving
    if overwrite:
        create_dirs(dst)
        data = df.to_dict('records')
        with open(dst, 'wb') as handle:
            pickle.dump(data, handle, protocol=pickle.HIGHEST_PROTOCOL)
        print(f'Saved at {dst}')

    return df

# example
# csv2pkl(src='../res/annotations/annotations_merged_balanced_renamed.csv', dst='../data/our_data/age_obj_cap_entries.pkl')

## pkl2csv

def pkl2csv(src, dst=None):
    if dst is None:
        dst = re.sub(r'\.pkl$', '.csv', src)
        
    # Loading
    with open(src, 'rb') as handle:
        obj = pickle.load(handle)
    df = pd.DataFrame(obj)
    
    # Overwrite?
    overwrite = True

    if os.path.exists(dst):
        while True:
            ans = str(input('A csv file already exists. Overwrite? [y]/n: ')).lower()
            if ans == 'y' or ans == 'yes':
                overwrite = True
                break
            elif ans == 'n' or ans == 'no':
                overwrite = False
                break

    # Saving
    if overwrite:
        create_dirs(dst)
        df.to_csv(dst)
        print(f'Saved at {dst}')
        
    return df

# example
# pkl2csv(src='../data/our_data/age_obj_cap_entries.pkl', dst='../data/our_data/age_obj_cap_entries.csv')

## Age dataset as pickle
In this section we make use of the previously defined functions to save age_dataset as a pickle file, to match the format required by the code.

csv2pkl(src='../data/age_data/age_dataset.csv')
